+++ urlParser/.idea/libraries/Maven__org_apache_maven_maven_compat_3_0_alpha_2.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.apache.maven:maven-compat:3.0-alpha-2">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/maven/maven-compat/3.0-alpha-2/maven-compat-3.0-alpha-2.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/maven/maven-compat/3.0-alpha-2/maven-compat-3.0-alpha-2-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/maven/maven-compat/3.0-alpha-2/maven-compat-3.0-alpha-2-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
