+++ urlParser/.idea/libraries/Maven__org_apache_maven_surefire_surefire_booter_3_0_0_M1.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.apache.maven.surefire:surefire-booter:3.0.0-M1">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/maven/surefire/surefire-booter/3.0.0-M1/surefire-booter-3.0.0-M1.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/maven/surefire/surefire-booter/3.0.0-M1/surefire-booter-3.0.0-M1-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/maven/surefire/surefire-booter/3.0.0-M1/surefire-booter-3.0.0-M1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
