+++ urlParser/.idea/libraries/Maven__org_codehaus_plexus_plexus_java_0_9_10.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.codehaus.plexus:plexus-java:0.9.10">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/codehaus/plexus/plexus-java/0.9.10/plexus-java-0.9.10.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/codehaus/plexus/plexus-java/0.9.10/plexus-java-0.9.10-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/codehaus/plexus/plexus-java/0.9.10/plexus-java-0.9.10-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
