+++ urlParser/pom.xml
@@ -38,6 +38,14 @@
             <type>maven-plugin</type>
         </dependency>
 
+        <!-- https://mvnrepository.com/artifact/org.apache.commons/commons-lang3 -->
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-lang3</artifactId>
+            <version>3.0</version>
+        </dependency>
+
+
     </dependencies>
 
     <!--<build>
